backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-dotenv